Fixed that the debugger window really moves itself to the foreground.
When the option "Pop into foreground when program stops" is turned on,
KDbg attempted to move itself into the foreground using XRaiseWindow().
This used to work with older windowmanagers, but it does no longer.
Quite some force is required to achieve the goal: KWin::raiseWindow() and
KWin::forceActiveWindow() are actually reserved for pager applications.
With some phantasy, the debugger event can be interpreted as "user interaction".

This fixes http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=171845