Do not annihilate redirections specified in the Program Arguments.
Unless "Full terminal emulation" was chosen for a program, the redirections
that can be specified in the Program Arguments were overridden by internal
redirections to /dev/null. With this modification the internal redirections
are placed first on the arguments line so that the user's redirections
come later and, therefore, prevail.

Reported by Christian Helbling.