Code cleanup: parsePrintExpr() need not return whether there was an error.
This indication was only used to avoid a call to inferTypesOfChildren(),
but this method is harmless if it is invoked for an error message. So we
just call it always if we get some parse result back.

Now that we don't need the return value for this indicator, we can use it for
the parse result.