Introduce a new object type that represents value trees of the parse results.
The debugger driver output parsers construct value trees, but used the
display item VarTree to represent the result. Now we use a new object type,
ExprValue, that is used to represent value trees. The purpose of this change
is to allow that the display items can be changed to a different type later.