Do not insert new values in reversed order.
Unfortunately, QListView inserts new items at the top instead of at the bottom.
When a new value is inserted at the top-most level, we must scan the list
for the last entry. Inserting new children is simpler, because we always
replace the entire list, so that we can keep track of the last one that
was inserted.