Maintain and highlight the current line.
Since the QTextEdit is used in read-only mode, no cursor is displayed.
This puts back the hightlight of the current line. It usually is the line
that also contains the (invisible) cursor of the QTextEdit, but there are
situations when this is not the case (but we do not care).