Restore the source window's context menu.
In the QTextEdit transition the context menu was replaced by the standard
context menu of the QTextEdit. This brings back our own context menu.

We must intercept the context menu events because the GUI factory, from
which we take the QPopupMenu object, does not relinquish ownership, as
would be required if we had overridden QTextEdit::createPopupMenu().