Insert a license note at the top of source files.
The code has always been licensed under the GPL v2 because this is what
the file COPYING said. Here we now make double-clear by stating this in
the source file.

We do not use the excessive license boilerplate that is customary
elsewhere because we do not think it is necessary to repeat the content
of COPYING.