Add primitive support for template types.
A type name in the type table that ends with '<*>' matches all
instantiated templates of that name, regardless of the actual template
parameter types.

With this tool, information in objects that does not depend on the actual
template parameter types can be shown, which we do with size and capacity
of std::vector.