Generalize template parameter patterns.
This implements a way to specify template parameters in the type tables
as patterns that are matched against the actual template instances.

A "*" in the place of a template parameter acts as a wild-card that
matches all types, but only one argument except that if the last
parameter is "*", then it matches all remaining template arguments.

The template pattern for std::vector is updated to take advantage of this
new scheme. The configuration section name in the .kdbgtt file for types
that contain a Template= pattern does not take part in the lookup, and
can, therefore, be an arbitrary name.