Add std::basic_string<char,...> and std::basic_ostringstream<char,...>.
If an automatic variable is of type std::string, then gdb says that it is
of type "string". However, if std::string is part of a nested type, then
gdb apparently cannot abbreviate the name to "string". With this addition
we can now display the string values of those nested values as well.

The situation with basic_ostringstream<> is slightly different: If an
automatic variable or parameter is defined with the typedefed name,
ostringstream, than gdb reports this short name; but if the template
instance name is used, gdb reports that.