Add QTL classes to qt3.kdbgtt.
This makes debugging those classes more convenient.

The node count of QValueList deserves some explanation: The node count
appears in a private class derived from QShared, and it could actually
be accessed like this:

  Expr1=(%s).sh->nodes

However, since the class is private, it is not always laid down in the
debugging information, and gdb doesn't know about it. Therefore, we
look up the member by treating the data pointed to by %s.sh as an array
of int. A complication arises because there is a node pointer between
.count and .nodes, and we must be prepared that an int is not the same
size as a pointer. Therefore, sizeof(char*)/sizeof(int) computes how many
ints there are in a pointer.