Pass the version number as a macro to the code.
With the automake based build system the version number was part of the
package specification in configure.in and was passed in the macro
'VERSION'. With cmake we get to choose our own macro name.

In order to avoid a complete rebuild when the version number changes, we
put the macro into its own file.