Do not store pointers to existing breakpoint objects in the command queues.
Previously, sometimes breakpoint commands were enqueued for existing
breakpoints. For consistency reasons, a pointer to the existing breakpoint
object was placed into the command object. Now we put the breakpoint ID
into the command object because we want to avoid that pointers to
existing breakpoint objects are "in flight". This allows us to rewrite
the breakpoint list container such that it holds the breakpoints as values
that can be relocated.