Fix incorrect format of the memory dump after switching expressions.
Consider the following use case:

1. Enter an expression A and change its format (call it format A).
2. Enter a new expression B. It is shown with format A.
3. Switch to expression A using the combobox list.
4. Change the format to A'.
5. Switch to expression B using the combobox list.

At this point expression B should be displayed with format A, but it is
shown with format A'. The problem was that the format was only placed into
the format cache if the switched-to expression is empty. We change this to
the opposite, i.e. if it is non-empty.