Fix dangerous substitution in a user-provided shell script.
A shell script is used to keep the external terminal window open. The user
can override it by (manually) specifying TermCmdStr in the rc file. (This
is only intended as an emergency.)

But this value is used as a format string for QString::sprintf(), so that
the first occurrence of "%s" is replaced by a file name. But if the
specified string contains more than one "%s" then a crash is very likely.

We now fix this by replacing all occurrences of "%s" (and only this
particular substring) by the file name.