Use QFileInfo to extract the last path component.
This implements SourceWindow::fileNameMatches() using QFileInfo. This
function checks whether the trailing file components match. It did so
using strcmp (!). This now uses QFileInfo objects to extract the last
component from the whole file name.

The title of the program settings dialog contains the name of the
debugged program. It was extracted manually, and we now use the services
of QFileInfo.