Show the active file's name as tab tooltip instead of in the title bar.
Since the source file's name is now visible in the tab, we do not need to
show it in the window title anymore, but we can show it as a tab tooltip.

Previously, the caption was updated every time the source file was
switched. This is no longer necessary, and it is sufficient to change the
title only when a new program is loaded.