Convert docking windows to Qt's QDockWindow.
This change is actually a regression because Qt's docking windows are
inferior to KDE's KDockWidget. But this brings the code closer to Qt4,
which has a full-featured QDockWidget.

Since in Qt4 the class is named QDockWidget, a typedef is introduced early
so that we can already write the code as if it were Qt4.