Rewrite infrastructure that searches for the prompt in the debugger output.
Previously, the search algorithm was implemented in the base class and was
parametrized by the derived classes. Rewrite this using a virtual function
so that the derived classes can implement specialized algorithms.

This removes a use of a regular expression that implied a conversion of
bytes to QString. The specialized algorithms can be written in a much
simpler manner than the generic algorithm.

The algorithm of XsldbgDriver is now slightly stricter because it accepts
the final characters of the prompt "> " only at the end of the output.