Remove unnecessary regular expressions.
The regular expression in parseMarker was searching for a constant string
at a fixed position, which is the same as strncmp(). Make it so.

The two regular expresssions in parseChangeExecutable() and parseCoreFile()
were only searching for a constant string anywhere in the text. Use
strstr() for this purpose.

In all cases an unnecessary implicit conversion of char* to QString was
removed.