Convert type tables.
The resource type is now added by referring to the existing resource type
"data", which is also used in CMakeLists.txt to install the type table
files.

The type table files themselves must be adjusted because the backslash is
an escape character that must be escaped to get a literal backslash for
the regular expressions.