Fix duplicated OK and Cancel buttons in two dialogs.
The Program Arguments and Attach to Process dialogs showed two sets of
OK and Cancel buttons because they derive from KDialog, which adds its
own buttons.

The solution is to use a regular QDialog as base class and add a
KDialogButtonBox in place of the explicit OK and Cancel buttons.

The change in the .ui files is mostly just re-indentation because the
top-most layout was removed, too.