Move layout code in breakpoint widget to a QtDesigner UI file.
This makes the layout easier to update and maintain in the future. It
removes the layout code from the logic in the rest of the breakpoint code,
making it all a bit smaller. Slots are renamed so that they are connected
automatically.

The code to compute a minimum size of the Enable/Disable button is removed
because the layout adjusts automatically if the button size must be
changed when the text is changed.

It also has a side effect of removing more legacy Qt3 classes and tidying
up any obsolete code with regards to layouts.