Update the breakpoint condition dialog to be a Qt Designer file
The class ConditionDlg now uses a .ui for setting up layout. The code has
been updated to match, with a lot of it being removed.

I have also updated the "ignore count" to be a spin box. This reduces some
code safely converting text to an integer, as well as being a lot clearer
to a user on how to use it.

The dialog has gained standard KDialog buttons, making it more consistent.

This also gets rid of a few legacy Q3 classes which were being used.