Convert source code window to Qt4's QPlainTextEdit.
Qt4 shows in examples/widgets/codeeditor how line numbers can be displayed
to the left of the source code lines. A simplified version of this is used
here (it does not change width depending on the magnitude of the line
numbers). The basic idea is to make the left margin large and to cover it
with a new widget, in which the extra information is displayed.

Special thanks to Gilles Allard, who provided an initial implementation.