Make ProgramTypeTable related TypeInfo infrastructure const-correct.
This is achieved by converting those Q3Dict to std::map that only keep
pointers. (Getting rid of Q3Dict is the main motivation for this change.)

This also fixes aliases: Previously, the aliased TypeInfo was registered
under the name of the referred-to name instead of the name of the alias.
Obviously, this has never worked correctly.