Fix program arguments dialog brown paper bag bug.
Since the conversion of the Boolean option list to QSet, invoking the
Program Arguments dialog crashed KDbg. A loop that is now at the outer
level was previously nested inside another loop (that was never entered when
no Boolean options were present at all) accesses a dialog control that is
not present.