Fix presentation of floating point registers.
Floating point registers are displayed by converting the raw byte sequence
to a long double value, which is then converted to a string representation
using QString::number(). However, this function does not have a long double
overload, hence, the value was always truncated to double before it was
displayed.

As Rick Regan noticed, the value -2.8309023271733244206437071197035129e-324
was displayed incorrectly as -4.94065645841246544177e-324.

This fixes it by using sprintf() instead of QString::number().