Treat continuation lines in scalar register values.
If GDB cannot print the cooked value of a scalar register on the current
line, it places is on the next line indented by a few spaces. This happens
particularly with the rip register when the address is decoded to a long
function name. Previously, the parsed register did not contain a cooked
register value, and the address on the contunation line was mis-parsed as
a register name. This fixes it.