Do not show new variables in red.
When the variable display was extended with a new entry, this new entry
was shown in red as if the value changed. The reason was that the entry
was allocated with no value, and subsequently, the entry was updated so
that it received the value that it ought to display. At this point, the
new value was different from the empty string and marked as changed.

Fix it by initializing the entry with its final value so that there is no
reason to mark it as changed.