Fix clicking the animation button.
Clicking the animation button should "Break" the running program. The
button's signal triggered(QAction*) was never emitted because no action is
associated with the button and the signal would be emitted only when the
action triggered. (What is this signal good for, by the way?) Connect to
signal clicked(bool) instead.