Show threads when attaching to a remote target.
Invoking kdbg with option -r allows to attach to a remote target. This
event should be treated as if the program had just stopped. In particular,
the list of threads should be filled. But it was forgotten, and this
change fixes it. Now the list of threads is available right away, and not
only at the next stop.