Ignore whitespace around warings.
Sometimes when printing variables, gdb says "warning: ..." twice and inserts
whitespace after the warning. Handle this.

For example:

(kdbg)print fn->asString()
warning: (Internal error: pc 0x8059d60 in read in psymtab, but not in symtab.)

warning: (Internal error: pc 0x8059d60 in read in psymtab, but not in symtab.)

$3 = { (... etc)