Fix thread list parsing for recent GDBs again.
Commit 449e95b9 (Fix parsing of the thread list) introduced a new algorithm
that can parse the thread list as produced by recent GDBs. But the new
method incorrectly skips two characters at the beginning of the stack
frame information, which strips off the first two letters of the function
name or the address.

The fix skips the two bytes only in the old parser, which are in that case
two blanks. These two are then trimmed when the name is constructed. This
incidentally also strips whitespace at the end of the name that was
inadvertently produced by the new algorithm.