Fix skipping an error message in parseBreakpoint().
The function receives the string to parse in variable 'output'. Its value
is copied to a local variable 'o', which is used to skip past an error
message. But later, variable 'output' is used.

This bug is present since day one, and not even the refactorings that
happened in 36acbde5 (Created an abstract base class DebuggerDriver, which
moved the function from brkpt.cpp to gdbdriver.cpp) and 261a7a51 (Fixed
watchpoints, which split off function parseNewBreakpoint()) did anything to
fix the bug. Fix it now.