Skip hints about duplicated breakpoints.
When a breakpoint is set at a location that already has a breakpoint, GDB
responds with "Note: breakpoints 3 and 4 also set at pc 0x41d5fa." It
confuses the parser, as if there was a syntax error. As a consequence,
KDbg treats the breakpoint as orphaned, but later it appears also in the
breakpoint list (because GDB did set the breakpoint), so that KDbg actually
keeps two records of the breakpoint. This eventually grows the breakpoint
list exponentially with each restart of the session.

Skip past such lines.

Noticed by Philip Ashmore.