Adjust breakpoint parser to a different syntax.
When a new breakpoint is set, GDB's response is used to seed the internal
table of breakpoints with file name and line number information. This is
in turn needed to display the new breakpoint in the source code window.

So far, we have expected that GDB's response is of this kind:

  Breakpoint 1 at 0x400b94: file multibrkpt.cpp, line 9.

But Luka Hranjec reported a case where the response looked like this:

  Breakpoint 4 at 0x804f158: lotto739.cpp:95.

Adjust the parser to accept this alternate syntax. Note that neither of
the two unambiguously identifies the file name and line number. Hence, we
just use the best-effort result that favors the former syntax.