Do not crash when the Memory window shows a dump after an error message.
When a memory dump of an expression is requested that refers to
inaccessible memory, an error message is displayed in the Memory window.
When the memory later becomes accessible, the memory dump should replace
the error message, but we crash instead due to an out-of-bounds access.
This fixes it.

The bug is most annoying because the address of the requested memory dump
is "sticky", and is not reverted by restarting the debug session; just
stepping over a memory allocation that makes the address accessible
triggered the bogus behavior, as was discovered by Terry Boldt.