Avoid incorrect length of logged commands.
The command that was just sent to the debugger is written to the log file.
The number of bytes to write was derived from the original Unicode string,
whereas the data was taken from the byte string (toLocal8Bit()). There is
no guarantee, though, that the length of the latter is the same as the
number of Unicode characters. Fix it by using the QIODevice::write()
overload that takes a QByteArray.