Ensure to communicate with the debugger in the C locale.
Modern GDBs offer translated messages. Set the default invocation to use
the C locale because that is the language in which its answers are
expected. We set LC_ALL because it overrides all other LC_* variables. It
should also override LANG, but let's unset it nevertheless, just in case.

Setting the language environment in this way also affects the programs
that are being debugged. The user can work it around by setting LC_ALL
from the Execution menu.