Keep pointer property of address values that are not strings.
Frequently, GDB prints an address followed by text in angle brackets. The
parser cleared the pointer property when it found something behind an
address, assuming that it is the value that the pointer points to. But
this assumption works only for strings.

In other cases, the text is something else, such as the name of the global
variable that the pointer value represents, e.g. 0x123456 <globalvar>.
Since the pointer property was cleared, the display did not allow to
expand such augmented addresses. Clear the pointer property only when we
know that there is a sting value.