Be prepared that the animated button is destroyed behind our back.
When the user changes the toolbar settings via the Configure Toolbar
dialog, the animated button is deleted by the framework because it is not
part of the settings. (It is re-instated the next time the application is
started.) But until then the pointer to the deleted object must be cleared.
A QPointer<> achieves just that. Fortunately, all accesses of the variable
do already check whether the pointer is valid.