Add add a 'make clean' target in the testprogs directory.
The testprogs directory is not under cmake control, but has its own
Makefile, because we do not need discovery of build flags and
dependencies. But to make it simple to rebuild all, 'make clean' would be
usefule. Add it. Note that all executables are built directly from the
source files without building object files explicitly. Therefore, only
the executables need to be removed.