Ignore request to jump to a negative frame index.
When the stack window is cleared, a currentRowChanged() signal is emitted,
which is connected to gotoFrame(). But then the item index is -1, and the
submitted command returns output showing no associated file information.
As a consequence, the frame pointer is cleared. The exact reason for this
change of behavior is not known. Most likely earlier versions of Qt did
not emit the signal when the list was cleared.