Trim line-breaks from GDB messages that go into the status bar.
Most notably, the message "Reading symbols from...done." that GDB writes
after loading an executable frequently contains an additional line-break
at the end. When it is shown in the status bar, it is formatted in two
lines, but there is not sufficient room, and so the line is truncated at
the top. Remove the trailing line-break.