Port KDbg to KF5.
Adjust CMakeLists.txt to the new world order of KF5.

This adds the icon to the application executable on Windows and MacOS.
However, the latter requires a 128px version, which is not present and
causes a warning (which we do not mind at this time).

Note, that the scalable icons are not installed because they should be
installed as *.svgz, but we do not want to place them under version
control in compressed form because any SVG changes are not presentable
as diffs.

A few source changes are also needed to compile for KF5.