Generate and install scalable icons.
We keep the scalable artwork as uncompressed SVG files because they can be
handled better under version control. In particular, file diffs can be
generated, which is a big plus.

But the installed versions of the artwork must be compressed to SVGZ. Do
this now during build time so that there is an installable version.