Use QDockWidget's toggle action to make docking windows visible.
The entries in the View menu do nothing. The reason is that the actions are
connected to slot of QDockWidget, but that does neither show nor hide the
Windows. Fortunately, QDockWidget offers an action object that can be used
instead.

On the flip side, we see now warnings on the command line because
KXMLGUIFactory does not like QAction objects so much (although things seem
to work).

Furthermore, the menu text is taken from the dock widget titles and cannot
be overridden in the action that is returned.