Request priority execution of commands using a new function.
In place of passing true for the clearLow argument of executeCmd(), add a
new function executeCmdOnce(). This will allow to implement the
executeCmd*() functions as template functions later. The trailing bool
parameter is in the way, and now that there is no caller that sets it
explicitly, we will remove it in the next commit.