Split queueCmd() into variants for each mode value.
We want to convert queueCmd() into a variadic template. The trailing mode
parameter is in the way. We want that call sites do not have to specify it.
For this purpose, introduce new variants with different names for the mode
values QMnormal and QMoverrideMoreEqual.

The original function is kept, but with a default argument, so that callers
do not have to specify it.

Adjust all callers to omit the mode argument and to call the new function
corresponding to the old mode argument.