Detect end of memory dump.
When the memory dump cannot be completed because the end of a virtual
memory segment is reached, GDB writes "Cannot access memory at address...".
This text is written literally into a cell of the memory dump window. This
is ugly because it makes one column wider than the others. Instead, detect
the case and just end the memory dump.