Avoid redundant memory dump requests when the combobox is dropped down.
After having requested a memory dump, the address expression is removed
from the combobox's list and inserted at the top. When the user drops down
the list later, the "current index" does not match the current text and an
signal activat() is sent. Avoid this by setting the current index to the
value inserted at the top of the list. Now the signal is sent only when a
text is actually selected.