Include the structure value in the value popup of the Expression window.
When a value shown in the Expression window is too long to be visible
entirely, hovering the mouse over the value shows a popup with the value.
However, it only shows the base value, not the value produced by the
structure inspectors. This is most inconvenient because the latter
contribute quite a bit to the text, but it is then truncated. Include the
structure values in the popup.