Avoid an overwrite confirmation when selecting a file argument.
The Program Arguments dialog has a button 'Insert file name' that permits
to select a file whose path name is then inserted in the program command
line edit box. Since the code uses getSaveFileName(), a warning "Do you
wish to overwrite [the file]?" is presented when an exisiting file is
selected.

We could use getOpenFileName() to avoid the warning, but then only existing
files can be selected. As we have to inject the option QFileDialog::AnyFile
in the dialog, we have to do it all manually.

While we are here, change the text of the button to "Insert", because it is
not an "Open" operation anymore.