Add basic support for changing GDB's disassembly flavor.
For x86 CPUs, GDB can print disassembly in either ATT format or Intel
format. Prepare the infrastructure that lets GDB know about a configured
flavor. After this change, it would be necessary to edit the program-
specific configuration file that can be found in the session files in
.local/share/kdbg/sessions. The setting is DisassemblyFlavor in group
General and can take the values 'att' and 'intel' (or nothing to leave
GDB's default setting).