Scroll output window as new text is added.
The output window should behave mostly like a terminal window that keeps
the most recent program output visible. There is an (invisible) cursor in
the window. Use ensureCursorVisible() to keep the cursor visible.

If the user wants to inhibit scrolling, they just can click in the window
somewhere. This will reposition the cursor and keep that location visible,
and not scroll to the end anymore. To resume scrolling, the user has to
click in the last line of the output window. Both actions may be difficult
to accomplish if the program writes output continuously, so the program is
best paused until the desired position is scrolled to and clicked on.