Be much more permissive with the output of the 'file' command.
GDB used to write a rather restricted set of sentences, but modern versions
are very chatty. The set of possible output lines can no longer anticipated
with reasonable effort. Therefore, accept everything after an initial line
that begins with "Reading symbols from" as success, and treat it as an
error if this line does not occur at the beginning of the output.