Recognize operator<=> within nested angle brackes.
Pointer variables can contain pointers to functions, where the function
name is listed after the pointer value:

   _M_p = 0x4106e3 <operator<=><...>(...)+66>

We have special logic to skip over the names of operators containing '<' or
'>'. However, we forgot that there is also operator<=>. In this case, the
final '>' was not treated, but closed a nesting level, leaving the
remaining text, typically a template or function parameter list, in the
stream that leads to a parse error. Make operator<=> known.