Do not mistake function template name ending in "operator" as operator.
Pointer variables can contain pointers to functions, where the function
name is listed after the pointer value:

   fn = 0x400d1f <anyoperator<int>(int const&)>

We have special logic to skip over the names of operators containing '<' or
'>'. However, when the function is a template whose name ends with
"operator", we mistake the part "operator<" as naming an operator and do
not introduce a new nesting level that would be needed to skip template
arguments in the intended way. Make sure that we recognize "operator" as
keyword only when it is not part of a longer name.