Heed program-specific debugger command even when run as root.
When KDbg is run as root, the user can opt to use or ignore a program-
specific debugger command. This check originates from the times when the
program-specific settings were stored next to the executable. In this case,
it would be possible for a user to construct a malicious debugger command,
and then entice a user who has sufficient privileges to debug the
executable with KDbg as root, which then could run the malicious command.

Since quite some time now, the program-specific settings are stored in the
user's home directory. The threat that this check wants to protect from
does not exist anymore. Remove the question whether to use the command when
run as root.