Require KFM 5.101.
KFM 5.101 has all features that will be needed to move on to KFM 6 and
warns about many constructs that will not be supported anymore.
Require this version so that we do not accidentally use old features that
happen to work only under KFM5.

To build the Debian packages, we have to advance to Ubuntu-24.04, because
version 22.04 is stuck at a version that is not recent enough. Thanks to
Sebastian Pipping for an initial patch.