Use QRegularExpression instead of QRegExp.
QRegExp updates its state with each match operation. Naturally, this is not
thread-safe. Enter QRegularExpression, which places captures in a new
object, QRegularExpressionMatch. This requires that all users of the new
class are updated accordingly.

The changes in regwnd.cpp are slightly more extensive, because exactMatch()
does not exist anymore. Use patterns that anchor to the beginning and end
of the subject string. Use the auxiliary function anchoredPattern() and
remove the anchors from the arguments.