Utilize a format string to reduce repeated concatenation of strings.
A filter string for a file dialog must have a particular pattern. Freeze it
in a format string instead of repeating the boilerplate parts.

It turns out that the file dialog does not mind the leading ";;", so we
leave it in for simplicity.

Since we have a QByteArray to QString conversion near the hash functions,
perform the replacement of the solidus in the QByteArray, because it is
easier to understand and avoids a narrow string to QString conversion.
Do the conversion of the QByteArray to QString in the return expression.